dummy-generator/requirements.txt
aio-pika==9.3.0
aiohttp==3.9.0
aiormq==6.7.7
aiosignal==1.3.1
amqp==5.2.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
autoflake==2.0.0
Babel==2.13.0
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
billiard==3.6.4.0
black==23.3.0
bleach==6.1.0
blis==0.7.11
catalogue==2.0.10
celery==5.2.7
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.4
confection==0.1.4
contourpy==1.1.1
cryptography==42.0.3
cycler==0.11.0
cymem==2.0.8
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.5
django-environ==0.11.2
django-filter==23.3
djangorestframework==3.14.0
dnspython==2.5.0
ecdsa==0.18.0
email-validator==1.3.0
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.78.0
fastjsonschema==2.18.1
flake8==6.0.0
flower==1.2.0
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
httpie==3.2.2
humanize==4.9.0
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
kafka-python==2.0.2
kiwisolver==1.4.5
kombu==5.3.5
langcodes==3.3.0
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.0.2
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.26.0
outcome==1.2.0
overrides==7.4.0
packaging==23.1
pamqp==3.2.1
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
password-validator==1.0
pathlib_abc==0.1.1
pathspec==0.12.1
pathy==0.11.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.6
psycopg2-binary==2.9.7
pure-eval==0.2.2
pyasn1==0.5.1
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.14
pyflakes==3.0.1
Pygments==2.16.1
pymongo==4.1.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.5
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0
pyzmq==25.1.1
referencing==0.30.2
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.3
rpds-py==0.10.6
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.2
seaborn==0.12.2
selenium==4.11.2
Send2Trash==1.8.2
six==1.16.0
sklearn==0.0.post9
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.2
starlette==0.19.1
terminado==0.17.1
thinc==8.1.12
threadpoolctl==3.2.0
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.66.2
traitlets==5.10.0
trio==0.22.2
trio-websocket==0.10.3
typer==0.7.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.18.2
vine==5.1.0
wasabi==0.10.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
whitenoise==6.5.0
wsproto==1.2.0
XlsxWriter==3.1.0
yarl==1.9.4
